add test to uninstaller

+++ unset-schedule.ps1
@@ -1,2 +1,6 @@
-Remove-Item "C:\CatiaKiller" -Force -Recurse
-Remove-Item "$env:userprofile\AppData\Roaming\Microsoft\Windows\Start Menu\Programs\Startup\launch.lnk" -Force
+if (Test-Path "C:\CatiaKiller"){
+    Remove-Item "C:\CatiaKiller" -Force -Recurse
+}
+if (Test-Path "$env:userprofile\AppData\Roaming\Microsoft\Windows\Start Menu\Programs\Startup\launch.lnk") {
+    Remove-Item "$env:userprofile\AppData\Roaming\Microsoft\Windows\Start Menu\Programs\Startup\launch.lnk" -Force
+}
